Refactor theme detection
