+++ src/main/java/com/tencao/saoui/SAOCore.kt
@@ -27,14 +27,11 @@ import com.tencao.saoui.config.FriendData
 import com.tencao.saoui.events.EventCore
 import com.tencao.saoui.screens.CoreGUI
 import com.tencao.saoui.themes.ThemeLoader
-import com.tencao.saoui.themes.elements.Hud
 import com.tencao.saoui.util.AdvancementUtil
 import com.tencao.saoui.util.DefaultStatsProvider
 import com.tencao.saoui.util.PlayerStats
-import com.tencao.saoui.util.append
 import net.minecraft.client.resources.SimpleReloadableResourceManager
 import net.minecraft.entity.EntityLivingBase
-import net.minecraft.util.ResourceLocation
 import net.minecraftforge.common.MinecraftForge
 import net.minecraftforge.fml.common.Loader
 import net.minecraftforge.fml.common.Mod
@@ -46,16 +43,7 @@ import net.minecraftforge.fml.common.eventhandler.EventBus
 import org.apache.logging.log4j.LogManager
 import org.apache.logging.log4j.Logger
 import java.io.File
-import java.io.IOException
-import java.nio.file.Files
-import java.nio.file.Path
-import java.nio.file.Paths
 import java.util.concurrent.ConcurrentHashMap
-import java.util.function.BiPredicate
-import java.util.stream.Stream
-import java.util.zip.ZipFile
-import javax.xml.bind.JAXBContext
-import kotlin.streams.toList
 
 @Mod(
     modid = SAOCore.MODID,
@@ -72,9 +60,9 @@ object SAOCore {
     const val VERSION = "2.1.7"
     const val DEPS = "required-after:saomclib@[1.4.8,);after:mantle"
 
-    val mc = Client.minecraft
+    val mc by lazy { Client.minecraft }
 
-    val saoConfDir: File = confDir(File(Client.minecraft.mcDataDir, "config"))
+    val saoConfDir: File by lazy { confDir(File(Client.minecraft.mcDataDir, "config")) }
     val isSAOMCLibServerSide: Boolean
         get() = SAOMCLib.proxy.isServerSideLoaded
 
@@ -135,98 +123,4 @@ object SAOCore {
     private fun confDir(genDir: File): File {
         return File(genDir, MODID)
     }
-
-    /**
-     * Returns a map of path to display name for themes
-     */
-    fun getFiles(): Map<ResourceLocation, ThemeMetadata> {
-        val candidates = mutableMapOf<ResourceLocation, ThemeMetadata>()
-
-        // Built-in
-        candidates += modFile.findThemePaths().extractThemesMetadata(MODID)
-
-        // Loaded resource packs
-        Client.minecraft.resourcePackRepository.repositoryEntries.map {
-            it.resourcePackName to Client.minecraft.resourcePackRepository.dirResourcepacks.resolve(it.resourcePackName)
-        }.forEach { (name, location) ->
-            candidates += location.findThemePaths().extractThemesMetadata(name)
-        }
-
-        val context = JAXBContext.newInstance(Hud::class.java)
-        val um = context.createUnmarshaller()
-        return candidates.filterValues {
-            try {
-                Client.resourceManager.getResource(it.themeRoot.append("/hud.xml")).inputStream.use { um.unmarshal(it) as Hud }
-                true
-            } catch (e: IOException) {
-                LOGGER.warn("Could not load HUD for $it !", e)
-                false
-            }
-        }
-    }
-
-    /**
-     * Find `hud.xml` files defined in given [File], which can be a zip/jar or folder
-     */
-    private fun File.findThemePaths(): List<Path> {
-        LOGGER.info("Analyzing $this")
-        return when {
-            // jar
-            this.isFile -> ZipFile(this).use { zip ->
-                zip.entries().asSequence()
-                    .filter { it.name.endsWith("hud.xml") && !it.isDirectory }
-                    .map { Paths.get(it.name) }
-                    .toList()
-            }
-            // folder
-            this.isDirectory -> {
-                val thisAsPath = this.toPath()
-                Files.find(thisAsPath, 5, BiPredicate { path, fileAttributes ->
-                    fileAttributes.isRegularFile && path.endsWith("hud.xml")
-                }).map { thisAsPath.relativize(it) }.use(Stream<Path>::toList)
-            }
-            // unknown
-            else -> {
-                LOGGER.warn("Unknown file type : $this")
-                emptyList()
-            }
-        }
-    }
-
-    /**
-     * Transform a list of paths into their respective metadata, indexed by id
-     * @param fallbackName Name to fall back to for legacy themes
-     */
-    private fun List<Path>.extractThemesMetadata(fallbackName: String): Map<ResourceLocation, ThemeMetadata> =
-        this.associate { path ->
-            val domain = path.subpath(1, 2).toString()
-            val name = path.parent.fileName.toString().let {
-                if (it != "themes") it else {
-                    LOGGER.warn("Theme pack $fallbackName is using the old theme structure !")
-                    fallbackName
-                }
-            }
-            val id = ResourceLocation(domain, name)
-            LOGGER.info("Read $id from $fallbackName at $path")
-            id to ThemeMetadata(
-                id = id,
-                themeRoot = ResourceLocation(domain, path.parent.toString().removePrefix("assets/$domain/")),
-                name = name.removeSuffix(".zip")
-            )
-        }
 }
-
-data class ThemeMetadata(val id: ResourceLocation, val themeRoot: ResourceLocation, val name: String)
-
-/*
-.map { path ->
-    val domain = path.subpath(1, 2).toString()
-    val name = path.parent.fileName.toString().takeUnless { it == "themes" } ?: domain
-    ResourceLocation(MODID, "themes/$name/hud.xml") to name
-    Triple(
-        ResourceLocation(domain, name),
-        ResourceLocation(domain, path.parent.toString()),
-        name
-    )
-}.asSequence().associateBy(Triple<*, *, *>::first)
- */
