+++ src/main/java/com/tencao/saoui/themes/ThemeMetadata.kt
@@ -1,0 +1,5 @@
+package com.tencao.saoui.themes
+
+import net.minecraft.util.ResourceLocation
+
+data class ThemeMetadata(val id: ResourceLocation, val themeRoot: ResourceLocation, val name: String)
